+++ apps/client/src/shared/blockly/tabConfig.ts
@@ -1,6 +1,5 @@
 import FixedFlyout from '@/core/fixedFlyout';
-import { htmlTagToolboxConfig } from './htmlTagToolboxConfig';
-import { cssStyleToolboxConfig } from './cssStyleToolboxConfig';
+import { cssStyleToolboxConfig, htmlTagToolboxConfig } from '@/shared/blockly';
 import StyleFlyout from '@/core/styleFlyout';
 import { TTabToolboxConfig } from '@/shared/types';
